+++ smart_dnn/tensor/TensorBackend.hpp
@@ -32,10 +32,10 @@ public:
     virtual Tensor scalarDiv(const double& scalar, const Tensor& tensor) const = 0;
 
     // Reduction operations
-    virtual Tensor sum(const Tensor& tensor, const std::vector<int>& axes, bool keepDims) const = 0;
-    virtual Tensor mean(const Tensor& tensor, const std::vector<int>& axes, bool keepDims) const = 0;
-    virtual Tensor max(const Tensor& tensor, const std::vector<int>& axes, bool keepDims) const = 0;
-    virtual Tensor min(const Tensor& tensor, const std::vector<int>& axes, bool keepDims) const = 0;
+    virtual Tensor sum(const Tensor& tensor, const std::vector<size_t>& axes, bool keepDims) const = 0;
+    virtual Tensor mean(const Tensor& tensor, const std::vector<size_t>& axes, bool keepDims) const = 0;
+    virtual Tensor max(const Tensor& tensor, const std::vector<size_t>& axes, bool keepDims) const = 0;
+    virtual Tensor min(const Tensor& tensor, const std::vector<size_t>& axes, bool keepDims) const = 0;
     virtual Tensor clip(const Tensor& tensor, const double& min, const double& max) const = 0;
 
     // Element-wise apply operations
